# internal/view/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Alumni API",
    "description": "İstanbul Üniversitesi mezun ağı. Kullanıcılar şimdilik bellekte tutuluyor, veritabanı henüz yok: sunucu yeniden başlayınca her şey sıfırlanır.",
    "version": "0.1.0"
  },
  "servers": [
    { "url": "http://localhost:8080", "description": "Yerel geliştirme" }
  ],
...
  "paths": {
...
    "/api/users/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "Kullanıcının id'si",
          "schema": { "type": "integer", "example": 1 }
        }
      ],
      "get": {
        "tags": ["ApiUserController"],
        "operationId": "ApiUserController.Show",
        "summary": "Tek bir kullanıcıyı getir",
        "responses": {
          "200": {
            "description": "Kullanıcı bulundu",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["ApiUserController"],
        "operationId": "ApiUserController.Update",
        "summary": "Kullanıcıyı bütünüyle değiştir",
        "description": "PUT kaydın yeni hâlinin tamamıdır, bu yüzden her alan zorunludur. Eksik alan gönderirsen 400 alırsın. Yalnızca bir alanı değiştirmek istiyorsan PATCH kullan.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewUser" },
              "example": { "name": "Ayşe Yıldız", "email": "ayse.yildiz@example.com", "department": "Bilgisayar Mühendisliği", "startYear": 2017, "graduationYear": 2021, "doubleMajor": "Matematik", "minor": "", "advisor": "Prof. Dr. Ahmet Kaya", "prepExempt": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Değiştirildi",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "patch": {
        "tags": ["ApiUserController"],
        "operationId": "ApiUserController.Patch",
        "summary": "Kullanıcının yalnızca gönderilen alanlarını değiştir",
        "description": "Gövdede anmadığın alanlar olduğu gibi kalır. En az bir alan göndermelisin.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UserPatch" },
              "example": { "graduationYear": 2022, "prepExempt": false }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Güncellendi",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["ApiUserController"],
        "operationId": "ApiUserController.Destroy",
        "summary": "Kullanıcıyı sil",
        "responses": {
          "204": { "description": "Silindi; geri gönderilecek gövde yok" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
    "/users/{id}/delete": {
      "parameters": [{ "$ref": "#/components/parameters/UserID" }],
      "post": {
        "tags": ["UserController"],
        "operationId": "UserController.Destroy",
        "summary": "Kullanıcıyı sil",
        "responses": {
          "303": { "$ref": "#/components/responses/RedirectToList" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/api/announcements": {
      "get": {
        "tags": ["ApiAnnouncementController"],
        "operationId": "ApiAnnouncementController.Index",
        "summary": "Bütün duyuruları listele (en yeni en üstte)",
        "responses": {
          "200": {
            "description": "Duyuru dizisi",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/Announcement" } }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["ApiAnnouncementController"],
        "operationId": "ApiAnnouncementController.Store",
        "summary": "Yeni duyuru oluştur",
        "description": "id ve createdAt alanlarını sunucu verir; gövdede gönderilirse yok sayılır.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewAnnouncement" },
              "example": { "title": "Kütüphane saatleri", "body": "Sınav haftası boyunca kütüphane 24 saat açık.", "author": "Kütüphane" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Oluşturuldu",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Announcement" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/api/announcements/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/AnnouncementID" }],
      "get": {
        "tags": ["ApiAnnouncementController"],
        "operationId": "ApiAnnouncementController.Show",
        "summary": "Tek bir duyuruyu getir",
        "responses": {
          "200": {
            "description": "Duyuru bulundu",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Announcement" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AnnouncementNotFound" }
        }
      },
      "put": {
        "tags": ["ApiAnnouncementController"],
        "operationId": "ApiAnnouncementController.Update",
        "summary": "Duyuruyu bütünüyle değiştir",
        "description": "Başlık, metin ve yapan kişi zorunludur. id ve createdAt değişmez.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/NewAnnouncement" },
              "example": { "title": "Mezunlar buluşması ertelendi", "body": "Buluşma 22 Kasım'a ertelendi.", "author": "Mezunlar Ofisi" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Değiştirildi",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Announcement" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AnnouncementNotFound" }
        }
      },
      "delete": {
        "tags": ["ApiAnnouncementController"],
        "operationId": "ApiAnnouncementController.Destroy",
        "summary": "Duyuruyu sil",
        "responses": {
          "204": { "description": "Silindi; geri gönderilecek gövde yok" },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AnnouncementNotFound" }
        }
      }
    },
    "/announcements": {
      "get": {
        "tags": ["AnnouncementController"],
        "operationId": "AnnouncementController.Index",
        "summary": "Duyuru listesi ve altında yeni duyuru formu",
        "responses": { "200": { "$ref": "#/components/responses/HtmlPage" } }
      },
      "post": {
        "tags": ["AnnouncementController"],
        "operationId": "AnnouncementController.Store",
        "summary": "Formdan yeni duyuru oluştur",
        "requestBody": {
          "required": true,
          "content": { "application/x-www-form-urlencoded": { "schema": { "$ref": "#/components/schemas/NewAnnouncement" } } }
        },
        "responses": {
          "303": {
            "description": "Tamam; tarayıcı duyuru listesine (/announcements) gönderilir",
            "headers": { "Location": { "schema": { "type": "string", "example": "/announcements" } } }
          },
          "400": { "$ref": "#/components/responses/FormWithError" }
        }
      }
    },
...
    "/announcements/{id}/delete": {
      "parameters": [{ "$ref": "#/components/parameters/AnnouncementID" }],
      "post": {
        "tags": ["AnnouncementController"],
        "operationId": "AnnouncementController.Destroy",
        "summary": "Duyuruyu sil",
        "responses": {
          "303": {
            "description": "Silindi; tarayıcı duyuru listesine (/announcements) gönderilir",
            "headers": { "Location": { "schema": { "type": "string", "example": "/announcements" } } }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "404": { "$ref": "#/components/responses/AnnouncementNotFound" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "AnnouncementID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Duyurunun id'si",
        "schema": { "type": "integer", "example": 1 }
      },
      "UserID": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Kullanıcının id'si",
        "schema": { "type": "integer", "example": 1 }
      }
    },
    "schemas": {
...
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true, "example": 1 },
          "name": { "type": "string", "example": "Ayşe Yılmaz" },
          "email": { "type": "string", "example": "ayse@example.com" },
          "department": { "type": "string", "example": "Bilgisayar Mühendisliği" },
          "startYear": { "type": "integer", "minimum": 1900, "example": 2017, "description": "Başlangıç yılı; 1900 ile bu yıl arasında" },
          "graduationYear": { "type": "integer", "example": 2021, "description": "Mezuniyet yılı; başlangıç yılından önce ve gelecekte olamaz" },
          "doubleMajor": { "type": "string", "example": "Matematik", "description": "ÇAP bölümü; yoksa boş" },
          "minor": { "type": "string", "example": "", "description": "Yandal bölümü; yoksa boş" },
          "advisor": { "type": "string", "example": "Prof. Dr. Ahmet Kaya", "description": "Danışman; yoksa boş" },
          "prepExempt": { "type": "boolean", "example": true, "description": "İngilizce hazırlıktan muaf mı?" }
        }
      },
      "NewUser": {
        "type": "object",
        "required": ["name", "email", "department", "startYear", "graduationYear"],
        "properties": {
          "name": { "type": "string", "example": "Elif Kaya" },
          "email": { "type": "string", "example": "elif@example.com" },
          "department": { "type": "string", "example": "Bilgisayar Mühendisliği" },
          "startYear": { "type": "integer", "minimum": 1900, "example": 2017, "description": "Başlangıç yılı; 1900 ile bu yıl arasında" },
          "graduationYear": { "type": "integer", "example": 2021, "description": "Mezuniyet yılı; başlangıç yılından önce ve gelecekte olamaz" },
          "doubleMajor": { "type": "string", "example": "Matematik", "description": "ÇAP bölümü; yoksa boş" },
          "minor": { "type": "string", "example": "", "description": "Yandal bölümü; yoksa boş" },
          "advisor": { "type": "string", "example": "Prof. Dr. Ahmet Kaya", "description": "Danışman; yoksa boş" },
          "prepExempt": { "type": "boolean", "example": true, "description": "İngilizce hazırlıktan muaf mı?" }
        }
      },
      "UserPatch": {
        "type": "object",
        "minProperties": 1,
        "description": "Gönderilen alanlar değişir. Sonuç yine bütün kurallara uymalı; örneğin mezuniyet yılı başlangıç yılından önce olamaz.",
        "properties": {
          "name": { "type": "string", "example": "Elif Kaya" },
          "email": { "type": "string", "example": "yeni@example.com" },
          "department": { "type": "string", "example": "Bilgisayar Mühendisliği" },
          "startYear": { "type": "integer", "minimum": 1900, "example": 2017, "description": "Başlangıç yılı; 1900 ile bu yıl arasında" },
          "graduationYear": { "type": "integer", "example": 2021, "description": "Mezuniyet yılı; başlangıç yılından önce ve gelecekte olamaz" },
          "doubleMajor": { "type": "string", "example": "Matematik", "description": "ÇAP bölümü; yoksa boş" },
          "minor": { "type": "string", "example": "", "description": "Yandal bölümü; yoksa boş" },
          "advisor": { "type": "string", "example": "Prof. Dr. Ahmet Kaya", "description": "Danışman; yoksa boş" },
          "prepExempt": { "type": "boolean", "example": true, "description": "İngilizce hazırlıktan muaf mı?" }
        }
      },
      "Announcement": {
        "type": "object",
        "properties": {
          "id": { "type": "integer", "readOnly": true, "example": 1 },
          "title": { "type": "string", "example": "Mezunlar buluşması" },
          "body": { "type": "string", "example": "Bu yılki mezunlar buluşması 15 Kasım'da Beyazıt kampüsünde." },
          "author": { "type": "string", "example": "Mezunlar Ofisi" },
          "createdAt": { "type": "string", "format": "date-time", "readOnly": true, "example": "2026-10-01T10:00:00+03:00" }
        }
      },
      "NewAnnouncement": {
        "type": "object",
        "required": ["title", "body", "author"],
        "properties": {
          "title": { "type": "string", "example": "Kütüphane saatleri" },
          "body": { "type": "string", "example": "Sınav haftası boyunca kütüphane 24 saat açık." },
          "author": { "type": "string", "example": "Kütüphane" }
        }
      }
    },
    "responses": {
      "AnnouncementNotFound": {
        "description": "Bu id'ye sahip duyuru yok",
        "content": {
          "text/plain": { "schema": { "type": "string" }, "example": "no announcement with that id" }
        }
      },
      "BadRequest": {
        "description": "İstek hatalı: bozuk JSON, eksik alan, kurala uymayan yıl ya da sayı olmayan id",
        "content": {
          "text/plain": { "schema": { "type": "string" }, "example": "name is required" }
        }
      },
      "HtmlPage": {
        "description": "HTML sayfası",
        "content": { "text/html": { "schema": { "type": "string" } } }
      },
      "RedirectToList": {
        "description": "Tamam; tarayıcı kullanıcı listesine (/users) gönderilir",
        "headers": { "Location": { "schema": { "type": "string", "example": "/users" } } }
      },
      "FormWithError": {
        "description": "Eksik alan: form, hata mesajı ve girilen değerlerle yeniden gösterilir",
        "content": { "text/html": { "schema": { "type": "string" } } }
      },
      "NotFound": {
        "description": "Bu id'ye sahip kullanıcı yok",
        "content": {
          "text/plain": { "schema": { "type": "string" }, "example": "no user with that id" }
        }
      }
    }
  }
}